Create README.md
In this project I used HTML, CSS, JavaScript, PHP, MySQL.
This project is about a travel agency, which allows a user to log in or register. It allows the user to book a ticket, also to add a comment with a review about the agency.
You have a chance to see all types of cars you can rent through our agency. You can see 3 top destinations for our agency!